Bug 1297716: Fix Mac OS cross build on mozilla-beta. r=rillian a=NPOTB
When we build mozilla-beta, mozharness forces the use of the universal
mozconfig. In universal/mozconfig.common, the SDK path is hardcoded in
the option --with-macos-sdk. We check if we are under the Linux
environment and then set the option to the correct path.

We also add all make variables and build options necessary for cross
builds.